refactor: dto is now using modern hierarchy style
