+++ src/dto/chat_info.rs
